FAIR-2507 删除一些文件
