+++ src/app/boot/next.js
@@ -223,10 +223,9 @@ winit.initNext = function () {
 	var firstStageLoaded = function(){
 		var sourceList = [
 			"app/view/base/sourceLoaded.js",
-			"app/postMessage/",
 			"earn/client/app/net/login.js",
 			"chat/client/app/net/login.js",
-			"app/viewLogic/login.js"
+			"app/net/login.js"
 		];
 		util.loadDir(sourceList, flags, fm, suffixCfg, function (fileMap) {
 			console.log("firstStageLoaded success-----------------");
@@ -235,65 +234,12 @@ winit.initNext = function () {
 			pi_modules.commonjs.exports.relativeGet("chat/client/app/net/init").exports.registerRpcStruct(fm);
 			// 活动注册
 			pi_modules.commonjs.exports.relativeGet("earn/client/app/net/init").exports.registerRpcStruct(fm);
-			var WebViewManager = pi_modules.commonjs.exports.relativeGet("pi/browser/webview").exports.WebViewManager;
-			WebViewManager.addListenStage(function(stage){
-				console.log('WebViewManager stage',stage)
-				if(stage === "firstStage"){    // 第一阶段完成  可以注册监听
-					// TODO 在回调中加载剩余代码 并且注册监听已经完成
-					var callGetHomePageEnterData = pi_modules.commonjs.exports.relativeGet("app/middleLayer/wrap").exports.callGetHomePageEnterData;
-					console.log('callGetHomePageEnterData start-----------------')
-					callGetHomePageEnterData().then(function(res){
-						console.log('callGetHomePageEnterData-------------------')
-						fpFlags.homePageReady = true;
-						fpFlags.homePageData = res;
-						enterApp();
-					});
-				}
-			});
-			console.log("stage webview goReady");
-			WebViewManager.getReady("firstStage");   // 通知一阶段准备完毕
 			// 继续加载首页所需
 			loadWalletFirstPageSource();  //钱包
-			if(!pi_update.inApp){
-				vmLoaded();
-			}
 		}, function (r) {
 			alert("加载目录失败, " + r.error + ":" + r.reason);
 		}, dirProcess.handler);
 	}
-
-	// vm代码加载
-	var vmLoaded = ()=>{
-		util.loadDir([ "app/store/","app/remote/postWalletMessage.js"], flags, fm, undefined, function (fileMap) {
-			// store 数据初始化完成后才通知 如果数据初始化没完成就通知  有可能数据还没读出来钱包就已经取了数据  导致数据不一致
-			var initStore = pi_modules.commonjs.exports.relativeGet("app/store/memstore").exports.initStore;
-			initStore().then(() => {
-				var WebViewManager = pi_modules.commonjs.exports.relativeGet("pi/browser/webview").exports.WebViewManager;
-				WebViewManager.addListenStage(function(stage){
-					if(stage === "firstStage"){    // 第一阶段完成  
-						// TODO 在回调中加载剩余代码 并且注册监听已经完成
-						var postStoreLoadedMessage = pi_modules.commonjs.exports.relativeGet("app/remote/postWalletMessage").exports.postStoreLoadedMessage;
-						console.log("postStoreLoadedMessage start-----------------------");
-						postStoreLoadedMessage();
-						// 加载剩下资源
-						util.loadDir([ "app/remote/","app/core_common/","app/core_pc/","app/publicLib/"], flags, fm, undefined, function (fileMap) {
-							var postAllLoadedMessage = pi_modules.commonjs.exports.relativeGet("app/remote/postWalletMessage").exports.postAllLoadedMessage;
-							console.log("postAllLoadedMessage start-----------------------");
-							postAllLoadedMessage();
-						}, function (r) {
-							console.log("加载目录失败, " + r.url + ", " + r.error + ":" + r.reason);
-						}, function(){});
-					}
-				});
-				console.log("stage vm goReady");
-				WebViewManager.getReady("firstStage"); // 通知一阶段准备完毕
-			});
-			
-		}, function (r) {
-			console.log("加载目录失败, " + r.url + ", " + r.error + ":" + r.reason);
-		}, function(){});
-	}
-
 	
 	// 加载钱包首页所需资源
 	var loadWalletFirstPageSource = function () {
@@ -306,10 +252,7 @@ winit.initNext = function () {
 			"pi/ui/lang.js",
 			"pi/ui/lang.tpl",
 			"app/view/base/",
-			'app/components/topBar/',
-            'app/components1/blankDiv/',
-			"app/components1/loading/",
-			"app/components1/btn/",
+			"app/components1/btn/",  
 			"app/res/css/",
 			'earn/client/app/res/css/'
 		];
@@ -317,7 +260,6 @@ winit.initNext = function () {
 			var tab = util.loadCssRes(fileMap);
 			tab.timeout = 90000;
 			tab.release();
-			fpFlags.walletReady = true;
 			enterApp();
 		}, function (r) {
 			alert("加载目录失败, " + r.error + ":" + r.reason);
@@ -326,16 +268,14 @@ winit.initNext = function () {
 
 	// 全部所需资源下载完成,进入app,显示界面
 	var enterApp = function(){
-		if(fpFlags.homePageReady && fpFlags.walletReady){
-			// 加载根组件
-			var root = pi_modules.commonjs.exports.relativeGet("pi/ui/root").exports;
-			root.cfg.full = false; //PC模式
-			var index = pi_modules.commonjs.exports.relativeGet("app/view/base/main").exports;
-			index.run(fpFlags.homePageData,function () {
-				// 关闭读取界面
-				document.body.removeChild(document.getElementById('rcmj_loading_log'));
-			});
-		}
+		// 加载根组件
+		var root = pi_modules.commonjs.exports.relativeGet("pi/ui/root").exports;
+		root.cfg.full = false; //PC模式
+		var index = pi_modules.commonjs.exports.relativeGet("app/view/base/main").exports;
+		index.run(function () {
+			// 关闭读取界面
+			document.body.removeChild(document.getElementById('rcmj_loading_log'));
+		});
 	}
 
 	// 加载一些需要预加载的图片
