+++ src/app/view/base/entrance.ts
@@ -1,5 +1,5 @@
 import { Widget } from '../../../pi/widget/widget';
-import { callTouristLogin } from '../../middleLayer/wrap';
+import { touristLogin } from '../../net/login';
 import { popNewMessage } from '../../utils/tools';
 
 /**
@@ -10,7 +10,7 @@ export class Entrance extends Widget {
     public ok:() => void;
     // 游客登录
     public async touristLoginClick() {
-        callTouristLogin().then(res => {
+        touristLogin().then(res => {
             popNewMessage('登录成功');
             this.ok && this.ok();
         }).catch (err => {
